moves adapter into destination
