+++ dlt/destinations/weaviate/weaviate_adapter.py
