+++ tests/load/weaviate/test_pipeline.py
@@ -4,7 +4,7 @@ import dlt
 from dlt.common import json
 from dlt.common.schema.typing import TTableSchema
 
-from dlt.helpers.weaviate_helper import weaviate_adapter, VECTORIZE_HINT, TOKENIZATION_HINT
+from dlt.destinations.weaviate.weaviate_adapter import weaviate_adapter, VECTORIZE_HINT, TOKENIZATION_HINT
 from dlt.destinations.weaviate.weaviate_client import WeaviateClient
 
 from tests.pipeline.utils import assert_load_info
